+++ packages/embed-wallet/src/createContext.ts
@@ -3,8 +3,8 @@ import { Context } from './types';
 export const createContext = (...contexts: (Context | null | undefined)[]): Context => {
   const defaultContext: Context = {
     app: {
-      name: window.origin,
-      url: window.origin,
+      name: window.location.hostname,
+      url: window.location.origin,
     },
   };