+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### vNext
 
 - [DAV-2698] Add SDK package and built-in authentication
+- [DAV-2359] Add wallet context support
 
 ### v1.4.0